refactor(android): route host adb through composition
