+++ src/platforms/android/adb-executor.ts
@@ -19,7 +19,9 @@ export {
   resolveAndroidTextInjector,
   resolveAndroidTouchProvider,
   resolveScopedAndroidAdbBackgroundTransport,
+  runAndroidHostAdb,
   withAndroidAdbProvider,
+  withAndroidHostAdbTransport,
   type AndroidAdbExecutor,
   type AndroidAdbExecutorOptions,
   type AndroidAdbExecutorResult,
