+++ src/platforms/android/emulator-lifecycle.ts
@@ -3,8 +3,7 @@ import type { DeviceInfo } from '@agent-device/kernel/device';
 import { AppError, asAppError } from '@agent-device/kernel/errors';
 import type { ExecResult } from '../../utils/exec.ts';
 import { runCmdDetached, whichCmd } from '../../utils/exec.ts';
-import { runAndroidHostAdb } from '@agent-device/platform-android/adb-executor';
-import './adb-host-binding.ts';
+import { runAndroidHostAdb } from './adb-executor.ts';
 import { Deadline, retryWithPolicy } from '../../utils/retry.ts';
 import { sleep } from '../../utils/timeouts.ts';
 import { bootFailureHint, classifyBootFailure } from '../boot-diagnostics.ts';
