+++ src/sdk/limrun-runtime-dependencies.ts
@@ -3,7 +3,7 @@ import type { LimrunRuntimeDependencies } from '@agent-device/provider-limrun';
 // ProviderDeviceRuntime.getInteractor is synchronous, so this previously eager factory remains the
 // deliberate static edge; making it lazy would require a proxy interactor rather than this seam.
 import { createAndroidInteractor } from '../core/interactors/android.ts';
-import { runAndroidHostAdb } from '@agent-device/platform-android/adb-executor';
+import { runAndroidHostAdb } from '../platforms/android/adb-executor.ts';
 import { execFailureDetails, runCmd } from '../utils/exec.ts';
 import { readVersion } from '../utils/version.ts';
